Add support for --parallel flag for Django 1.9+
This required fixing how we get the total test count.

When running tests in parallel, a core suite is generated to
contain a collection of sub-suites that are then ran in parallel.

In that case we have to go and retrieve the count from each
of the sub-suites as the parent will not directly contain the tests.